UI Integration
Menu is now integrated with the backend code.
