+++ background/background.js
@@ -2,5 +2,5 @@
 browser.browserAction.onClicked.addListener( async function()
 {
     const tabs = await browser.tabs.query({active:true,currentWindow:true});
-    browser.tabs.create({url:"platforms/intro.html"});
+    browser.tabs.create({url:"platforms/menu.html"});
 })
