--- platforms/add.html
@@ -1,142 +1,0 @@
-<!DOCTYPE html>
-<style>
-body
-{
-    background-image: linear-gradient(rgb(112,61,214), rgb(188,196,255));
-    background-repeat: no-repeat;
-    background-attachment:fixed;
-    color: rgb(44, 11, 111);
-    text-align:center;
-}
-input[type=text]
-{
-    width: 100;
-    padding: 12px 20px;
-    display: inline-block;
-    border: 1px solid #ccc;
-    box-sizing: border-box;
-    border-radius: 4px;
-}
-.plusButton
-{
-    font-weight:bold;
-    font-size:30;
-}
-.modal {
-  display: none;
-  position: fixed; 
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%; 
-  overflow: auto; 
-  background-color: rgb(0,0,0); 
-  background-color: rgba(0,0,0,0.4); 
-  -webkit-animation-name: fadeIn; 
-  -webkit-animation-duration: 0.4s;
-  animation-name: fadeIn;
-  animation-duration: 0.4s
-}
-
-/* Modal Content */
-.modal-content {
-  position: fixed;
-  bottom: 0;
-  background-color: #fefefe;
-  width: 100%;
-  -webkit-animation-name: slideIn;
-  -webkit-animation-duration: 0.4s;
-  animation-name: slideIn;
-  animation-duration: 0.4s
-}
-
-/* The Close Button */
-.close {
-  color: white;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-}
-
-.close:hover,
-.close:focus {
-  color: #000;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.modal-header {
-  padding: 2px 16px;
-  background-color:rgb(90, 41, 187);
-  color: white;
-}
-
-.modal-body {padding: 2px 16px;}
-
-.modal-footer {
-  padding: 2px 16px;
-  background-color: rgb(90, 41, 187);
-  color: white;
-}
-
-/* Add Animation */
-@-webkit-keyframes slideIn {
-  from {bottom: -300px; opacity: 0} 
-  to {bottom: 0; opacity: 1}
-}
-
-@keyframes slideIn {
-  from {bottom: -300px; opacity: 0}
-  to {bottom: 0; opacity: 1}
-}
-
-@-webkit-keyframes fadeIn {
-  from {opacity: 0} 
-  to {opacity: 1}
-}
-
-@keyframes fadeIn {
-  from {opacity: 0} 
-  to {opacity: 1}
-}
-</style>
-<script type="module" src ="../content-scripts/serverConnection.js"></script>
-<script src="../content-scripts/addLink.js" type="module"></script>
-<script src="../content-scripts/aws-min.js"></script>
-<head>
-    <title>GroupLink</title>
-</head>
-<body>
-    <div id="link"class="link">
-    <h1> GroupLink Extension</h1>
-    <h2>Enter the links that you would like to assemble into a group link.</h2>
-    <input class="linkChoice" type = "text" placeholder="Enter link"></input>
-    </div>
-    <br>
-    <button id="plusButton"class="plusButton">+</button>
-    <br>
-    <br>
-    <input type="text" class="nameLink" placeholder="Enter a name for the link"></input>
-    <br>
-    <br>
-    <input type="checkbox" id="newWindow">
-    <label for="newWindow">Open in a new window?</label>
-    <br>
-    <br>
-    <button onclick = "saveinStorage()" id="submitButton">Create Your Group Link!</button>
-    <div id="submitModal" class="modal">
-        <div class="modal-content">
-          <div class="modal-header">
-            <span class="close">&times;</span>
-            <h2>Link Created!</h2>
-          </div>
-          <div class="modal-body">
-            <a id="submitText"></a>
-          </div>
-          <div class="modal-footer">
-            <h3>Click the link above to view your GroupLink.</h3>
-          </div>
-        </div>
-    </div>      
-</body>
