--- platforms/intro.html
@@ -1,43 +1,0 @@
-<style>
-body
-{
-    background-image: linear-gradient(rgb(112,61,214), rgb(188,196,255));
-    background-repeat: no-repeat;
-    background-attachment:fixed;
-    color: white;
-    text-align:center;
-}
-a
-{
-    height:120;
-    width:250;
-    background-image: linear-gradient(rgb(112,61,214), rgb(188,196,255));
-    color:white;
-    border: 4px solid white;
-    transition: all 0.6s;
-    font-size:20;
-}
-a:hover
-{
-    height:140;
-    width:270;
-    background-image: linear-gradient(rgb(188,196,255), rgb(112,61,214));
-    color:white;
-    border: 4px solid white;
-    transition: all 0.6s;
-    font-size: 25;
-}
-</style>
-<head>
-    <title>Group Link Introduction</title>
-</head>
-<body>
-    <h1>Welcome to Group Link</h1>
-    <h2>With GroupLink, saving closely related links has never been so simple.</h2>
-    <br>
-    <br>
-    <a href="add.html">Create a New Group Link</a>
-    <br>
-    <br>
-    <a href="view.html">View Group Links Saved In Storage</a>
-</body>
