+++ platforms/menu.css
@@ -1,0 +1,214 @@
+body 
+{
+  background-color: #e1b4f9;
+}
+div 
+{
+  text-align: left;
+}
+hr 
+{
+  height: 10px;
+}
+.navbar 
+{
+  margin: 10px;
+}
+.navbar label 
+{
+  border-bottom: 3px solid white;
+  display: inline;
+  font-size: 25px;
+  font-family: "Gill Sans", sans-serif;
+  font-weight: bold;
+  padding: 10px;
+  transition: all 0.6s;
+}
+.navbar label:hover 
+{
+  border-bottom: 3px solid black;
+  transition: all 0.6s;
+}
+.selectionContent label,
+.linkChoices label,
+.choiceAndSubmit label {
+  font-size: 18px;
+  font-family: "Gill Sans", sans-serif;
+  font-weight: bold;
+  padding: 7px;
+}
+.selectionContent,
+.linkMenu,
+.links,
+.profile {
+  color: white;
+  background-color: #c085d6;
+  line-height: 25px;
+}
+.GroupLinkName,
+.linkDisplay {
+  width: fit-content;
+  font-weight: bold;
+}
+.center-screen {
+  flex-flow: row wrap;
+}
+.center-child {
+  position: absolute;
+  top: 0%;
+  left: 50%;
+  transform: translate(-50%, 20%);
+}
+.nameTitle {
+  font-size: 50px;
+}
+input[type="text"] {
+  height: 30px;
+  width: 200px;
+  background-color: #f6e2fa;
+}
+input[type=checkbox] {
+  height: 20px;
+  width: 20px;
+  outline: 2px solid black;
+}
+.plusButton {
+  background-color: #3c2747;
+  color: white;
+  font-weight: bold;
+  border-radius: 4px;
+  border: #3c2747;
+  font-size: 18px;
+  text-align: right;
+}
+.submitButton {
+  padding: 10px;
+  background-color: #3c2747;
+  color: white;
+  font-weight: bold;
+  font-size: 20px;
+  border-radius: 12px;
+  border: #3c2747;
+  margin: 10px;
+}
+.choiceAndSubmit {
+  text-align: center;
+}
+.linkMenu {
+  display: block;
+}
+.links {
+  display: none;
+}
+
+.link,
+.nameLink {
+  border: #d8b7f4;
+  margin: 10px;
+}
+
+.modal {
+  display: none;
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgb(0, 0, 0);
+  background-color: rgba(0, 0, 0, 0.4);
+  -webkit-animation-name: fadeIn;
+  -webkit-animation-duration: 0.4s;
+  animation-name: fadeIn;
+  animation-duration: 0.4s;
+}
+
+/* Modal Content */
+.modal-content {
+  position: fixed;
+  bottom: 0;
+  background-color: #fefefe;
+  width: 100%;
+  -webkit-animation-name: slideIn;
+  -webkit-animation-duration: 0.4s;
+  animation-name: slideIn;
+  animation-duration: 0.4s;
+}
+
+/* The Close Button */
+.close {
+  color: white;
+  float: right;
+  font-size: 28px;
+  font-weight: bold;
+}
+
+.close:hover,
+.close:focus {
+  color: #000;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.modal-header {
+  padding: 2px 16px;
+  background-color: #3c2747;
+  color: white;
+}
+
+.modal-body {
+  padding: 2px 16px;
+}
+
+.modal-footer {
+  padding: 2px 16px;
+  background-color: #3c2747;
+  color: white;
+}
+
+/* Add Animation */
+@-webkit-keyframes slideIn {
+  from {
+    bottom: -300px;
+    opacity: 0;
+  }
+  to {
+    bottom: 0;
+    opacity: 1;
+  }
+}
+
+@keyframes slideIn {
+  from {
+    bottom: -300px;
+    opacity: 0;
+  }
+  to {
+    bottom: 0;
+    opacity: 1;
+  }
+}
+
+@-webkit-keyframes fadeIn 
+{
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+.selectionButton 
+{
+  width: 100%;
+}
