+++ platforms/menu.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<head>
+    <title>GroupLink</title>
+</head>
+<body>
+    <script src="../content-scripts/aws-min.js"></script>
+    <script type="module" src="../content-scripts/serverConnection.js"></script>
+    <link rel="stylesheet" href="menu.css">
+    <div class="center-screen center-child">
+        <h1 class="nameTitle">GroupLink</h1>
+        <div class="navbar">
+            <label id="createLink">Create Link</label>
+            <label id="savedLinks">Saved Links</label>
+            <label id="profile">Profile</label>
+        </div>
+        <div id = "linkMenu" class="linkMenu">
+            <div class="selectionContent">
+                <div class="GroupLinkName">
+                    <br>
+                    <label>GroupLink name:</label>
+                    <br>
+                    <input class="nameLink" type="text">
+                </div>
+                <br>
+            </div>
+            <div class="linkChoices">
+                <div class="linkDisplay">
+                    <label class="displayTitle">Links to be grouped:</label>
+                    <button id="plusButton" class="plusButton displayTitle">+</button>
+                </div>
+                <br>
+                <input type="text" size = "90" class="link" style="width:80%;">
+            </div>
+            <br>
+            <br>
+            <br>
+            <br>
+            <br>
+            <div class="choiceAndSubmit">
+                <input id="newWindow" type="checkbox">
+                <label>Have GroupLink open in new window</label>
+                <br>
+                <button id="submitButton" class="submitButton">Create GroupLink</button>
+            </div>
+        </div>
+        <div id="links" class="links">
+        </div>
+    </div>
+    <script type="module" src="../content-scripts/menu.js"></script>
+    <script src="../content-scripts/savedLinks.js"></script>
+    <div id="submitModal" class="modal">
+        <div class="modal-content">
+          <div class="modal-header">
+            <span class="close">&times;</span>
+            <h2>Link Created!</h2>
+          </div>
+          <div class="modal-body">
+            <a id="submitText"></a>
+          </div>
+          <div class="modal-footer">
+            <h3>Click the link above to view your GroupLink.</h3>
+          </div>
+        </div>
+    </div>
+</body>
