--- platforms/package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "aws-sdk": "^2.1232.0",
-    "fs": "^0.0.1-security",
-    "fs-js": "^1.0.6"
-  }
-}
