--- platforms/test.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<style>
-body
-{
-    background-image: linear-gradient(rgb(112,61,214), rgb(188,196,255));
-    background-repeat: no-repeat;
-    background-attachment:fixed;
-    color: white;
-    text-align:center;
-}
-input[type=text]
-{
-    width: 100;
-    padding: 12px 20px;
-    display: inline-block;
-    border: 1px solid #ccc;
-    box-sizing: border-box;
-    border-radius: 4px;
-}
-.plusButton
-{
-    font-weight:bold;
-    font-size:30;
-}
-</style>
-<script src="../content-scripts/addLink.js" type="module"></script>
-<head>
-    <title>GroupLink</title>
-</head>
-<body>
-    <div id="link"class="link">
-    <h1> GroupLink Extension</h1>
-    <h2>Enter the links that you would like to assemble into a group link.</h2>
-    <input class="linkChoice" type = "text" placeholder="Enter link"></input>
-    </div>
-    <br>
-    <button id="plusButton"class="plusButton">+</button>
-    <br>
-    <br>
-    <input type="text" class="nameLink" placeholder="Enter a name for the link"></input>
-    <br>
-    <br>
-    <input type="checkbox" id="newWindow">
-    <label for="newWindow">Open in a new window?</label>
-    <br>
-    <br>
-    <button id="submitButton">Create Your Group Link!</button>
-</body>
