--- platforms/view.html
@@ -1,33 +1,0 @@
-<style>
-    body
-{
-    background-image: linear-gradient(rgb(112,61,214), rgb(188,196,255));
-    background-repeat: no-repeat;
-    background-attachment:fixed;
-    color: white;
-    text-align:center;
-}
-.selectionButton
-{
-    font-size: 20;
-    transition: all 0.6s;
-    font-weight: normal;
-    border: 1px solid white;
-    transition: all 0.6s;
-}
-.selectionButton:hover
-{
-    border: 3px solid purple;
-    transition: border 0.6s;
-}
-</style>
-<head>
-    <title>GroupLink Currently Saved Links</title>
-</head>
-<body>
-    <h1>GroupLinks Saved in Storage (save these before closing your browser)</h1>
-    <div id="savedLinks">
-        <p>List of Links:</p>
-    </div>
-    <script src="../content-scripts/savedLinks.js"></script>
-</body>
